+++ pcs/lib/pacemaker/values.py
@@ -23,6 +23,7 @@ _ID_FIRST_CHAR_NOT_RE = re.compile("[^a-zA-Z_]")
 _ID_REST_CHARS_NOT_RE = re.compile("[^a-zA-Z0-9_.-]")
 SCORE_INFINITY = "INFINITY"
 
+
 def __exec(name):
     return os.path.join(settings.pacemaker_binaries, name)
 
@@ -68,11 +69,7 @@ def is_score(value: str) -> bool:
 
 
 def is_duration(runner: CommandRunner, value: str) -> bool:
-    cmd = [
-        __exec("iso8601"),
-        "--duration",
-        value
-    ]
+    cmd = [__exec("iso8601"), "--duration", value]
     _, _, retval = runner.run(cmd)
     return retval == 0